requirement.txt
streamlit
Pillow
matplotlib
numpy
opencv-python
opencv-python-headless